requirements.txt
pandas==1.3.5
Pillow==9.4.0
pyspark==3.3.2
tensorflow==2.11.0
ipython